test(core): check more
