+++ tests/integration/test_services.py
@@ -134,58 +134,3 @@ def test_video_loader_from_batch():
     assert results[1] == 15  # Keep using max batch size where it can
     assert results[2] == 5  # Final 5 frames in the 50 frame video
     assert len(results) == 3  # total num batches
-
-
-def test_get_directory_listing():
-    """Tests the accociated function that it generates the correct jsTree json."""
-    result = services.get_directory_listing(
-        "tests/integration/test_data/test_directory_listing_folder"
-    )
-    assert result == [
-        {
-            "id": "tests/integration/test_data/test_directory_listing_folder",
-            "text": "test_directory_listing_folder",
-            "type": "folder",
-            "children": [
-                {
-                    "id": "tests/integration/test_data/test_directory_listing_folder/folderA",
-                    "text": "folderA",
-                    "type": "folder",
-                    "children": True,
-                },
-                {
-                    "id": "tests/integration/test_data/test_directory_listing_folder/folderB",
-                    "text": "folderB",
-                    "type": "folder",
-                    "children": True,
-                },
-                {
-                    "id": "tests/integration/test_data/test_directory_listing_folder/emptyfile",
-                    "text": "emptyfile",
-                    "type": "file",
-                },
-                {
-                    "id": "tests/integration/test_data/test_directory_listing_folder/invalidext.in21p3",
-                    "text": "invalidext.in21p3",
-                    "type": "file",
-                },
-                {
-                    "id": "tests/integration/test_data/test_directory_listing_folder/text.txt",
-                    "text": "text.txt",
-                    "type": "text/plain",
-                },
-                {
-                    "id": "tests/integration/test_data/test_directory_listing_folder/video.mp4",
-                    "text": "video.mp4",
-                    "type": "video/mp4",
-                },
-            ],
-        }
-    ]
-    with pytest.raises(NotADirectoryError):
-        services.get_directory_listing(
-            "tests/integration/test_data/test_directory_listing_folder/emptyfile"
-        )
-
-    with pytest.raises(FileNotFoundError):
-        services.get_directory_listing("this/path/should/not/exist")
